test: Add statetest with contract, move existing integration tests
