+++ x/evm/testutil/StateTest.json
@@ -1,0 +1,4 @@
+{
+  "abi": "[{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"tempChangeEmpty\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"tempChangeNonEmpty\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"value1\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"value2\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]",
+  "bin": "608060405261123460015534801561001657600080fd5b50610422806100266000396000f3fe608060405234801561001057600080fd5b506004361061004c5760003560e01c80633033413b1461005157806349a8fee01461006f5780635d33a27f1461008b578063893dd4d1146100a9575b600080fd5b6100596100c5565b604051610066919061021e565b60405180910390f35b6100896004803603810190610084919061026a565b6100cb565b005b610093610165565b6040516100a0919061021e565b60405180910390f35b6100c360048036038101906100be919061026a565b61016b565b005b60005481565b600080541461010f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610106906102f4565b60405180910390fd5b600060015403610154576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161014b90610360565b60405180910390fd5b806000819055506000808190555050565b60015481565b600081036101ae576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016101a5906103cc565b60405180910390fd5b6000600154036101f3576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016101ea906103cc565b60405180910390fd5b60006001819055508060018190555050565b6000819050919050565b61021881610205565b82525050565b6000602082019050610233600083018461020f565b92915050565b600080fd5b61024781610205565b811461025257600080fd5b50565b6000813590506102648161023e565b92915050565b6000602082840312156102805761027f610239565b5b600061028e84828501610255565b91505092915050565b600082825260208201905092915050565b7f56616c7565206973206e6f7420656d7074790000000000000000000000000000600082015250565b60006102de601283610297565b91506102e9826102a8565b602082019050919050565b6000602082019050818103600083015261030d816102d1565b9050919050565b7f4e65772076616c756520656d7074790000000000000000000000000000000000600082015250565b600061034a600f83610297565b915061035582610314565b602082019050919050565b600060208201905081810360008301526103798161033d565b9050919050565b7f56616c756520697320656d707479000000000000000000000000000000000000600082015250565b60006103b6600e83610297565b91506103c182610380565b602082019050919050565b600060208201905081810360008301526103e5816103a9565b905091905056fea264697066735822122033275f5cb76a3770398b9c120d76b7979041e44f0eae668a1124fb413468762764736f6c63430008180033"
+}
